[CI] Remove windows build from Circle
Upgrading Electron necessitates that we remove the codebase's dependency on prebuilt
binaries. The prebuilt binary repo we have relied on is stale and will break Windows
and Linux builds when re-compilation is automatically attempted by node-gyp rebuild.

This commit removes Windows from the Circle CI config as we will begin building for
Windows in a native environment.